+++ mbb_ppt/__init__.py
@@ -22,4 +22,4 @@ from .review import SlideReviewer, AutoFixPipeline, review, autofix
 
 ExecEngine = MbbEngine
 
-__version__ = '0.2.2'
+__version__ = '0.3.0'